update: 紙吹雪の雨のアニメーションのリファクタリング
